Add support for 32bit and 64bit.

+++ zplugin-dds/src/dds_mgt.rs
@@ -71,7 +71,7 @@ unsafe extern "C" fn on_data(dr: dds_entity_t, arg: *mut std::os::raw::c_void) {
         dr,
         samples.as_mut_ptr() as *mut *mut raw::c_void,
         si.as_mut_ptr() as *mut dds_sample_info_t,
-        MAX_SAMPLES as u64,
+        MAX_SAMPLES as usize,
         MAX_SAMPLES as u32,
     );
     for i in 0..n {
